StockApi developments-4.

+++ src/main/java/com/ndr/stockexchanger/service/StockService.java
@@ -1,11 +1,17 @@
 package com.ndr.stockexchanger.service;
 
+import java.time.ZonedDateTime;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.ndr.stockexchanger.api.dto.PriceUpdateRequestDTO;
 import com.ndr.stockexchanger.api.dto.StockDTO;
 import com.ndr.stockexchanger.domain.Stock;
 import com.ndr.stockexchanger.repository.StockRepository;
+
+import jakarta.validation.Valid;
 
 @Service
 public class StockService {
@@ -21,18 +27,40 @@ public class StockService {
 		stock.setCurrentPrice(requestStockDto.getCurrentPrice());
 		stock.setLastUpdate(requestStockDto.getLastUpdate());
 		stock = stockRepository.save(stock);
-		StockDTO responseStockDTO = new StockDTO();
-		responseStockDTO.setId(stock.getId());
-		responseStockDTO.setName(stock.getName());
-		responseStockDTO.setDescription(stock.getDescription());
-		responseStockDTO.setCurrentPrice(stock.getCurrentPrice());
-		responseStockDTO.setLastUpdate(stock.getLastUpdate());
-		return responseStockDTO;
+		
+		return generateStockDTO(stock);
 	}
 
 
-	public void deleteStock(Long stockId) {
+	public boolean deleteStock(Long stockId) {
+		Optional<Stock> optStock = stockRepository.findById(stockId);
+		if (optStock.isEmpty())
+			return false;
 		stockRepository.deleteById(stockId);
+		return true;
+	}
+
+
+	public StockDTO updatePrice(Long stockId, @Valid PriceUpdateRequestDTO priceUpdateRequest) {
+		Optional<Stock> optStock = stockRepository.findById(stockId);
+		if (optStock.isEmpty())
+			return null;
+		Stock updateStock = optStock.get();
+		updateStock.setCurrentPrice(priceUpdateRequest.getNewPrice());
+		updateStock.setLastUpdate(ZonedDateTime.now());
+		updateStock = stockRepository.save(updateStock);
+		return generateStockDTO(updateStock);
+	}
+
+
+	private StockDTO generateStockDTO(Stock stock) {
+		StockDTO stockDTO = new StockDTO();
+		stockDTO.setId(stock.getId());
+		stockDTO.setName(stock.getName());
+		stockDTO.setDescription(stock.getDescription());
+		stockDTO.setCurrentPrice(stock.getCurrentPrice());
+		stockDTO.setLastUpdate(stock.getLastUpdate());
+		return stockDTO;
 	}
 
 }
